Added solar volt support
Update the water level screen
Fix bug in sendUpdate when ADS1115 not available

+++ Agrismart/include/watering.hpp
@@ -2,6 +2,6 @@
 #define WATERING_HPP_
 bool turnOnTOF(bool display_text = true);
 String waterLevelPercentage(bool VL53L0X_alive);
-unsigned int soilMoisture();
+float soilMoisture();
 void pumpWater(unsigned int powerInPercentage);
 #endif
